requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
autopep8==1.5.7
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.54.0
django-plotly-dash==2.2.2
djangorestframework==3.12.4
docopt==0.6.2
dpd-static-support==0.0.5
dpd_components==0.1.0
Flask==3.0.2
fonttools==4.48.1
frozenlist==1.4.1
gunicorn==20.1.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
multidict==6.0.5
num2words==0.5.13
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plotly==5.18.0
pluggy==1.4.0
psycopg2-binary==2.9.7
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.2
stripe==8.2.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.6.0
toml==0.10.2
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
Unipath==1.1
urllib3==2.2.0
Werkzeug==3.0.1
whitenoise==5.3.0
yarl==1.9.4
zipp==3.17.0
